Releasing 0.6.0
Skipping a test that segfaults only on 1.5.0 and 2.0.0.  The test was
test_NR_DEC_issue188_beach_64bitsbox_jp2_41_decode.  It was not introduced
until the 2.1.x series and the fix was backported to 2.0.1.

glymur.test.test_jp2k.TestJp2k.test_no_cxform_pclr_jpx was failing on
1.5.2 only.  It's a pretty obscure test, so just skipping it on that
release.

Some tests for warnings are being skipped.