remove test requirement for six
This requirement is problematic, as some versions cause problems with
assertWarns and some do not.  We are already filtering out versions that
we know cause problems and it can unintentionally cause an upgrade
superceding the system version if it remains as a "test_requires"
option, so best to get rid of it.