remove test requirement for six, doc modifications, doctest fix
The test requirement for six was problematic, as some versions cause
problems with assertWarns and some do not.  We are already filtering
out versions that we know cause problems and it can unintentionally
cause an upgrade superceding the system version if it remains as a
"test_requires" option, so best to get rid of it.  Additionally, the
"test_requires" option was entirely removed as pip doesn't recognize it
upon install, so "mock" would not be installed on Python2.7 via pip.

The doctest fix was the real fix for #316, not sure why this did
not show up when testing on linux mint