Relax installation requirements
On Ubuntu 12.04.5 LTS, the system supplies these packages:

Package    Ubuntu 12.04.5 LTS   Glymur requirements
--------   ------------------   -------------------
openjpeg        1.3.0                 1.5.0
numpy:          1.6.1                 1.7.0
mock:           0.7.2                 1.0.1
lxml:           2.3.2                 3.0.0

The ubuntu-packaged version of numpy is definitely too old; the stated
numpy requirement is correct.  The system version of openjpeg is also
too old.

But if we install numpy via pip and source-compile version 2.1.0 of
openjpeg, the tests pass with the system versions of mock and lxml, so
we can relax the requirements for these two packages.

One test that did not work was
test_NR_ENC_random_issue_0005_tif_12_encode from
glymur.test.test_opj_suite_write.TestSuiteWrite.   The
problem there was that matplotlib at the time ('1.1.1rc')
seems to use the old PIL package to try to read certain
image formats, and it doesn't handle this particular TIFF
correctly.  If we detect this configuration, we skip the test.

Closes-Issue: #333