Release candidate for 0.8.3
Fix the following bugs:

If the config file is missing the [library] section, then it should
be completely ignored.  No other sections are currently being used,
but they may be in a future release.

Closes-Issue: #342

Skip gdal info test on cygwin.  The difference may be due to a build
difference in gdal.  It may take a considerable amount of investigation
to determine this, and it isn't important enough to delay 0.8.3.

References-Issue:  #345

Fix typo in windows installation instructions.
Simplify the travis install steps.
Fix test-on-install bug where degenerate geotiff test file not found.
Fix README.md for mention of python 3.5