Remove most of OPJ_DATA_ROOT requirement
So much reliance on OPJ_DATA_ROOT is problematic, as most of the tests
will not pass in Travis-CI.  The tests have been rewritten so as to
remove this requirement.

Many new warnings added that subclass UserWarning

TLM segment constructor altered to be consistent with other segment
constructors, i.e. length and offset should come last.

Changed a RuntimeError to DifferingSubsampleFactorsError

Closes-Issue: #349