Fix test exclusion when no library is found
The version string in this case is '0.0.0' which was not handled by the
regular expression in the test suite.

Skip warning tests on python2

It seems that we just can't rely upon warning verification in python2.

Fix test exclusion when no library is found.  The version string
in this case is '0.0.0' which was not handled by the regular
expression in the test suite.

Closes-Issue: #366

Spotcheck warning messages.