Remove WARNING_INFRASTRUCTURE test fixtures
This was introduced when certain versions of the "six" package are in
play (versions 1.0-1.6).  It made it difficult to test on such
platforms.  See issues #223, #255, #259, #260, and #314.  Since six is
not a core part of the language and is only needed for testing, and
since six has progressed far beyond 1.7 (currently at 1.10), we probably
don't need to worry about this anymore.  And it just cluttered up the
testing code.

Closes-Issue: #370