Rework detection and handling of UTF-8 BOM
The XMLBox parsing wasn't handling the BOM as well as it should
have.  Both python2 and python3 can decode the UTF-8 BOM just fine
since we force decoding with the utf-8 codec.  The problem comes
when __str__ is invoked, because it uses the ascii codec on python2.
So we really only need to remove the BOM on python2.

The existing test for this was incorrectly written due to the ascii
codec on python2.

Closes-Issue:  #356