Streamline XML processing with encoding decls
It isn't necessary to remove an encoding declaration, then read the text
in as an Element, then reconstitute as an ElementTree.  We can just use
the ElementTree.parse method on a bytesio wrapper around the text.  That
way we do not have to worry about the encoding declaration at all.

Closes-Issue: #388