Improve code coverage
Move glymur.test to tests as that subpackage does not need to be
installed.

Move ICCProfile code to its own module.

Move glymur.lib.tests into tests, mv _ICCProfile class into own module.

Remove code for Exif Photo and Interop IFDS as it never worked and
wasn't even hooked in.

Drop support for openjpeg 2.0.x

Closes-Issue: 397

Disable ICC profile printing test on 2.7.  Printing always seems
to be different here for different versions of python and it's not
worth trying to track it.  Going forward, we should just track the
most recent version of python.

Closes-Issue: 399

Remove ignore_pclr_cmap_cdef from keywords

In the old read method, we say we allow for passing  ignore_pclr_cmap_cdef,
but it cannot be passed into the _read function.

Closes-Issue: #400

Fix repr for BitsPerComponentBox

Original code wasn't even close to what it needed to be.  Missing
arguments for BPC and SIGNED.

Fix parsing error for FreeBox

Was only reading LT section of the free box, not reading the dead
payload part.

Closes-Issue: #401

Fix repr of UnknownBox

Need to quote the box ID, otherwise the ID is looked up as a variable.

Closes-Issue: #403