Fix test to run on CentOS 6 with openjpeg 1.3
The test tests.test_jp2k.TestJp2k_1_x.test_read_version_15 was failing
on a CentOS 6.8 machine that had openjpeg 2.1.2 installed through
anaconda, but also had openjpeg 1.3.0 installed through the system
package manager.  Version 1.3 is too old to run the test, but not so old
that the openjpeg module can't load it.