Do not allow UnknownBox to have a valid identifier
If there is an UnknownBox, it usually comes about by being a corrupt
version of a known box with a known box identifier.  Allowing that known
identifier to be associated with a corrupt box can lead to confusing
situations.  For example, a PCLR box is crucial to determining the
"shape" property of a Jp2k object, but if the PCLR box is corrupt, then
the shape calculation cannot rely on it.  Seems like the most
straightforward way to deal with this is to just use a generic invalid
identifier for the unknown box, such as xxxx.