Remove use of tempdir.NamedTemporaryFile
The test suite cannot easily run tests that use
tempfile.NamedTemporaryFile because of issue14243 at bugs.python.org.
Replace this pattern by using a base class that sets up a test directory,
provides several filenames in that directory that are ready for use,
and removes the directory after each test.