Rework the marker parsing function table
All segments with markers not specifically mentioned in the
marker parsing function table are now handled as if they
are reserved segments.

Some functions are renamed to reflect the fact that they work on markers
rather than marker segments.

Remove _VALID_MARKERS list, it is unused.

If a codestream is truncated between markers, the exception raised is now
struct.error instead of IOError.