Update for Python 3.8
Update CI setup and docs to mention 3.8 support.  Restrict gdal to be
less than 3.0.0.

Always prefer conda-forge to the default channel.  By prepending the
conda-forge channel, hopefully we can get a more across-the-board
conda-forge build.  As it is now, we get a lot of packages from the
default channel, and there may be room for incompatibilities there.

Remove python 3.4 artifacts. Build and test was already largely gone,
but a few test fixtures tied to python 3.4 remained.  Mention of 3.4
removed from setup.py.

Remove python2 constructs from py2 skipped test.  The test was already
marked to skip on python2, so no reason to keep any python2 code.