Print geotiff UUID as EXIF if gdal errors out
In the case referenced in this issue, an XML sidecar has the necessary
information.  Unfortunately, there does not seem to be a straightforward
way to gain access to the XML file in a general way.  The best way
forward seems to be to fall back to printing the GEOTIFF header raw keys
in case gdal errors out.

GDAL version 3.0 has recently come out, and it/proj will print the
information differently depending on the respective versions, so we must
account for that.

Fedora 31: gdal 2.3.2 and proj 5.2.0
MacPorts:  gdal 3.0.1 and proj 6.2.1
Anaconda:  gdal 2.3.3 and proj 5.2.0