Rewrite Travis-CI interface
There was an issue with conda-forge and poppler that could not be
resolved and was resulting in failures on all but python 3.8.  Since the
python versions < 3.8 did not actually need conda-forge, we can go back
to using the defaults channel, but 3.8 still needs conda-forge.

The 3.6 builds need importlib_resources, but this is not available
through defaults channel, so we need to use pip for this.  This means
removing the requirements*.txt files and going full-on with YAML
configuration files.  This necessitates a big rewrite.

As a final insult, gdal just recently moved to 3.0.2 in defaults, which
breaks the geotiff test again.  We will require gdal<3.0 for now.

Remove one unneeded shell script from ci.