Remove bad library location test points
If the configuration file points to a bad library location, the result
is different on Windows than on mac/linux.  On mac/linux, there is a
warning.  On windows, there is an OSError raised, with a pop-up window
hindering the test point.

There were two very similar test points for this.

Since this essentially testing python and not glymur, these test points
should be removed.