Skip config file test on windows
The test for having a config file in the current directory is
problematic on windows.  In order to really test this, we need to create
an entry for the library that is named differently than the "system"
library path name.  This can be done on mac/linux easily with a symlink,
but this is problematic on windows as it seems to involve elevated
privileges.  Same with just copying the library.  Again, since config
files are getting a bit deprecated, I think it's ok to just skip this on
windows.