Fix fragile config file test
The testpoint
tests.test_config.TestSuiteConfigFile.test_config_file_in_current_directory
is even more fragile than thought. It will fail on Linux in the case of
anaconda unless openjpeg is also installed via anaconda. If libopenjp2
is loaded via ctypes, we don't get the path information back.

Instead of doing gymnastics to deal with paths to libraries that can't
be determined reliably on linux, the best way forward is to dial back on
the expectations of the test.  Just verify that the code to read a
configuration file in the current directory works.

Resolves: #509