Add support for openjpeg 2.4.0
Adjust one test.  On openjpeg versions < 2.4.0, one test tried to
retrieve an image with a 0x0 component, requiring special logic to
detect the situation.  In version 2.4.0, the library detects this.
The test now issues a different exception because of this.

Add test matrix for openjpeg 2.4.  No support in conda-forge for openjpeg
> 2.3, gdal, and scikit-image all together just yet.

References: #525