Do not use TIFFGetFieldDefaulted
This API call can be problematic if the tag has more that one possible
datatype on a platform that is hard to test, say a big endian platform.
Get around this by using the IFD as read via programming to the TIFF
spec.

Refactor, pass image width, height to readRGBAImage instead of using
TIFFGetFieldDefaulted.

Closes: #545