Fix geotiff tests
The test_print_bad_geotiff test was being skipped due to the fog of
war.  The test was re-enabled and a warning is now being verified due to
the GeoTIFF header being corrupt.

The test__printing__geotiff_uuid__xml_sidecar now warns due to the
above changes.

Skip two tests on windows, to be fixed in issue560.