Add options for supporting ResolutionBoxes
Add capture_resolution and display_resolution keyword parameters
to Jp2k and Tiff2Jp2k init/constructors.

There was a fair amount of desperately needed refactoring of the
Jp2k __init__ method.

The _write_superbox method was rewritten due to problems inherent to
opening a JP2 file in append mode and having to rewind the file pointer
to fill in the superbox length.

Closes: #563