Rewrite superbox code
Still now sure why, but appending a superbox onto the end of a JP2 file
was not working, the file pointer was not trustworth when we would re-seek
to write the length of the superbox.  Get around this by writing entirely
to a buffer, then append the completed buffer to the JP2 file.