Improve handling of technically invalid JP2 files
Files with multiple jp2 header boxes are invalid but are still readable
by openjpeg.  It is not necessary to error out.

Issue a warning in case there are multiple JP2H boxes.  Do not continue
to error out is there are multiple JP2C boxes, but issue a warning
instead.  Continue to error out if there are no JP2C boxes.