Stop using libtiff to create test files
There's an issue with libtiff on mac arm64 that results in test
failures.  It seems that libtiff doesn't want to take planarconfig
values via TIFFSetTag thru ctypes.  Very odd.  Better to create the
files once on another platform and then refer to the files as test
fixtures.

Two sets of files, one set derived from scikit-image datafiles, and one
set that is not.

Closes: #593