Catch negative decoded components earlier
Negative decoded components are not allowed.  This used to be caught as
an OpenJPEGLibraryError, but the error message incorrectly gives the component
number as the roll-over from a negative number to a very large uint32.
This is starting to raise a numpy warning that will eventually result in
an error, so we now catch it before it reaches either numpy or the
openjpeg library.  The exception is now a ValueError.

Closes: #598