Refactor Jp2k class into a reader and a writer
This hopefully simplifies things a bit.  The new class is Jp2kReader.
The Jp2k class remains as the 'writer' and is a subclass of Jp2kReader.

The only backwards incompatibility introduced concerns the shape
paroperty.  The shape.setter has been removed since it no longer makes
sense to do that except in the constructor.

Closes: #issue620