Do not error out trying to read JPH with Jp2k
The root cause here was that the opj codec was being reset in Jp2k after
already being set in Jp2kr, which caused glymur to try to reparse the
JPH file as if it were J2K.  This was because the file type was not
recognized.  So change Jp2k to not reset the codec if it has already
been set and allow for some basic recognition of JPH features, if only
to not issue unnecessary warnings.

Closes: #660