Fix number-of-tile-rows computation
When converting a TIFF to JP2, the number of rows of tiles in the output
JP2 file was incorrectly computed.  This error would manifest itself
when the image was not "square" in terms of number of tile rows and
colums, and when the last row of JP2 tiles were partial.

Closes: #665